# openapi/swagger.yaml
openapi: 3.1.0
info:
  title: HRMS Elite API v1
  version: "1.0.0"
  description: |
    OpenAPI specification for HRMS Elite API version 1.
    Includes authentication, document management and user endpoints.
servers:
  - url: http://localhost:3000
    description: Development server
paths:
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                companyId:
                  type: string
      responses:
        '200':
          description: Logged in successfully
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '200':
          description: Logged out successfully
  /api/v1/auth/user:
    get:
      tags: [Authentication]
      summary: Get current user
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '200':
          description: Current user returned
...
components:
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session authentication cookie
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF protection token
...